Replace any active untitled untouched document on Open (#257)
This may be a controversial opinion.

It felt weird to open edit, open a file, and then see "FILE + 1". I
didn't care about the empty Untitled document, and there was no reason
for it to be kept around.

After this pull request, if you launch edit and then Open a new document
without editing the existing placeholder untitled document or giving it
a name, Edit will close that document and open the new one.

The same applies to New: if you press New, and then Open, Edit will
close the untitled temporary document.